refactors parse_filings and makes some adjustments
