+++ .gitignore
@@ -9,3 +9,4 @@ geckodriver.log
 result.csv
 result.json
 cases.db*
+.DS_Store
